Improve UI

+++ src/state.rs
@@ -45,15 +45,14 @@ pub struct Unit {
     /// The unit identifier
     pub id: UnitId,
 
-    // TODO: Support negative coordinates
-    /// X/Y position in the world in mm
+    /// X/Y position in the world in m
     pub position: [f64; 2],
 
-    /// Angle of the unit in milli degrees
+    /// Angle of the unit in radiant
     pub angle: f64,
 
     /// Direction and speed of the movement. The angle of the movement may be different then the
-    /// angle of the unit. The unit is mm per milli second.
+    /// angle of the unit. The unit is m per milli second.
     pub speed_vector: [f64; 2],
 
     /// Health of the unit
@@ -116,14 +115,14 @@ impl GameState {
 /// This needs to be transferred to the client only once, on connecting.
 #[derive(Serialize, Deserialize, PartialEq, Debug, Clone)]
 pub struct WorldState {
-    /// Width of the world in mm
-    pub x: u64,
-    /// Height of the world in mm
-    pub y: u64,
+    /// Width of the world in m
+    pub x: f64,
+    /// Height of the world in m
+    pub y: f64,
 }
 
 impl WorldState {
-    pub fn new(x: u64, y: u64) -> WorldState {
+    pub fn new(x: f64, y: f64) -> WorldState {
         WorldState {
             x: x,
             y: y,
